remove useless indirection
